pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aquadet"
version = "0.1.0"
description = "Edge AI system for aquatic waste detection, sizing, tracking, and telemetry"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "AquaDet Team" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "numpy>=1.26.0",
  "opencv-python>=4.10.0",
  "torch>=2.3.0",
  "torchvision>=0.18.0",
  "pyyaml>=6.0.1",
  "onnx>=1.16.0",
  "onnxruntime>=1.18.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
train = [
  "ultralytics>=8.3.0",
  "albumentations>=1.4.12",
  "tensorboard>=2.17.0",
  "scikit-learn>=1.5.1"
]
export = [
  "tensorrt; platform_system == 'Linux'",
  "pycuda; platform_system == 'Linux'"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_core*", "backend*", "edge*", "iot*"]
